+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Add ``IsTrue``.
 - Add ``IsFalse``.
+- Make ``Except`` **not** call *value* if it's callable. (**breaking change**)
 
 
 v0.0.1 (2015-05-07)